move rdf function to new file, merge master
